+++ mod/racing_plus/PostEffectInit.lua
@@ -15,16 +15,6 @@ function PostEffectInit:Effect15(effect)
   end
 end
 
--- EffectVariant.FART (34)
-function PostEffectInit:Effect34(effect)
-  if g.run.changeFartColor == true then
-    -- We want special rolls to have a different fart color to distinguish them
-    g.run.changeFartColor = false
-    local color = Color(5.5, 0.2, 0.2, 1, 0, 0, 0) -- Bright red
-    effect:SetColor(color, 0, 0, false, false)
-  end
-end
-
 -- EffectVariant.HOT_BOMB_FIRE (51)
 function PostEffectInit:Effect51(effect)
   -- Turn enemy fires into a different color